refactor: 경매상세페이지에서만 쓰이는 컴포넌트 내부로 이동
