+++ src/app/auctions/[auctionId]/_component/AuctionDetailInfo/DefaultCase.tsx
