+++ src/app/auctions/[auctionId]/_component/AuctionDetailInfo/DefaultComponent.tsx
