+++ src/app/auctions/[auctionId]/_component/AuctionDetailInfo/index.tsx
