+++ src/app/auctions/[auctionId]/_component/LineChart/index.tsx
